+++ ci/.travis_test.sh
@@ -23,6 +23,7 @@ if [[ "$COVERAGE" == "true" ]]; then
     nosetests -s -v --with-coverage --cover-package=mlxtend.regression
     nosetests -s -v --with-coverage --cover-package=mlxtend.sklearn
     nosetests -s -v --with-coverage --cover-package=mlxtend.text
+    nosetests -s -v --with-coverage --cover-package=mlxtend.feature_selection
 else
     nosetests -s -v mlxtend.classifier
     nosetests -s -v mlxtend.evaluate
@@ -31,5 +32,6 @@ else
     nosetests -s -v mlxtend.regression
     nosetests -s -v mlxtend.sklearn
     nosetests -s -v mlxtend.text
+	nosetests -s -v mlxtend.feature_selection
 fi
 #make test-doc test-sphinxext