+++ mlxtend/classifier/__init__.py
@@ -5,4 +5,7 @@ from .perceptron import Perceptron
 from .adaline import Adaline
 from .logistic_regression import LogisticRegression
 from .neuralnet_mlp import NeuralNetMLP
-__all__ = ["Perceptron", "Adaline", "LogisticRegression", "NeuralNetMLP"]
+from .ensemble import EnsembleClassifier
+
+__all__ = ["Perceptron", "Adaline", "LogisticRegression", "NeuralNetMLP",
+           "EnsembleClassifier"]