+++ mlxtend/feature_selection/__init__.py
@@ -1,0 +1,8 @@
+# Sebastian Raschka 08/19/2014
+# mlxtend Machine Learning Library Extensions
+# Submodules with preprocessing functions.
+
+from .feature_sel import ColumnSelector
+from .sequential_backward_select import SBS
+
+__all__ = ["ColumnSelector", "SBS"]