+++ mlxtend/feature_selection/feature_sel.py
@@ -2,15 +2,15 @@
 # mlxtend Machine Learning Library Extensions
 # scikit-learn utilities for feature selection
 
+
 class ColumnSelector(object):
     """ A feature selector for scikit-learn's Pipeline class that returns
         specified columns from a numpy array.
-    
+
     """
-    
     def __init__(self, cols):
         self.cols = cols
-        
+
     def transform(self, X, y=None):
         return X[:, self.cols]