+++ mlxtend/preprocessing/__init__.py
@@ -2,10 +2,13 @@
 # mlxtend Machine Learning Library Extensions
 # Submodules with preprocessing functions.
 
+from .transformer import TransformerObj
 from .mean_centering import MeanCenterer
 from .shuffle import shuffle_arrays_unison
 from .scaling import minmax_scaling
 from .scaling import standardizing
+from .dense_transformer import DenseTransformer
+
 
 __all__ = ["MeanCenterer", "shuffle_arrays_unison",
-           "minmax_scaling", "standardizing"]
+           "minmax_scaling", "standardizing", "DenseTransformer"]