+++ mlxtend/preprocessing/transformer.py
@@ -7,7 +7,7 @@ import numpy as np
 class TransformerObj(object):
     def __init__(self):
         self.ary = None
-        
+
     def _get_array(self, X):
         if isinstance(X, list):
             self.ary = np.asarray(X)