+++ tests/tests_preprocessing/test_shuffle_shuffle_arrays_unison.py
@@ -1,6 +1,7 @@
 import numpy as np
 from mlxtend.preprocessing import shuffle_arrays_unison
 
+
 def test_shuffle_arrays_unison():
     X1 = np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
     y1 = np.array([1, 2, 3])