+++ mlxtend/__init__.py
@@ -1,5 +1,4 @@
 # Sebastian Raschka 2014-2015
 # mlxtend Machine Learning Library Extensions
 
-__version__ = '0.2.8'
-
+__version__ = '0.2.9'